Bugfixes for feature extractor behaviour on Linux and custom path robustness
BugFix: Set the default Rendering Hardware Interface (RHI) for Linux to OpenGL4.3 instead of Vulkan.  Using VULKAN RHI causes the Depth and ClassSegmentation feature extractor output to be black.

BugFix: When specifying directories (e.g. materials, textures), NDDS will work correctly irrespective of the user adding a trailing slash or not to the end of the path.